LayoutTests/media/media-source/media-source-webm.html

<!DOCTYPE html>
<html>
<head>
    <title>media-source-webm</title>
    <script src="media-source-loader.js"></script>
    <script src="../video-test.js"></script>
    <script>
    var loader;
    var source;
    var sourceBuffer;
    var resizeEventFired = false;
    var videoWidth = 0;
    var videoHeight = 0;

    function loaderPromise(loader) {
        return new Promise((resolve, reject) => {
            loader.onload = resolve;
            loader.onerror = reject;
        });
    }

    function resizePromise(video, width, height) {
        return new Promise(resolve => {
            video.addEventListener('resize', event => {
                // First resize can be 0x0
                if (video.videoWidth == 0 && video.videoHeight == 0)
                    return;
                resizeEventFired = true;
                videoWidth = video.videoWidth;
                videoHeight = video.videoHeight;
                resolve();
            });
        });
    }

    window.addEventListener('load', async event => {
        try {
            findMediaElement();
            loader = new MediaSourceLoader('content/test-vp9-manifest.json');
            await loaderPromise(loader);

            source = new MediaSource();
            run('video.src = URL.createObjectURL(source)');
            await waitFor(source, 'sourceopen');
            waitForEventAndFail('error');

            run('source.duration = loader.duration()');
            run('sourceBuffer = source.addSourceBuffer(loader.type())');
            const resized = resizePromise(video);
            run('sourceBuffer.appendBuffer(loader.initSegment())');

            await waitFor(sourceBuffer, 'update');

            consoleWrite('Append a media segment.')
            run('sourceBuffer.appendBuffer(loader.mediaSegment(0))');

            await resized;
            testExpected('resizeEventFired', true);
            testExpected('videoWidth', 320);
            testExpected('videoHeight', 240);
            endTest();
        } catch (e) {
            failTest(`Caught exception: "${e}"`);
        }
    });
    </script>
</head>
<body>
    <div>
        This tests the ability of the SourceBuffer to reset the parser after an abort(). A SourceBuffer in this state should be able to accept
        a new initialization segment or a new media segment.
    </div>
    <video controls></video>
</body>
</html>
